update status values

+++ firebase_db.py
@@ -81,11 +81,12 @@ def add_history(history):
     history_ref.push(history)
     history_max = round(datetime.now(timezone.utc).timestamp()) - (3600*4)      # 4hrs of history
     #snapshot = history_ref.order_by_key().limit_to_last(1).get()
-    snapshot = history_ref.order_by_child("time").end_at(history_max).get()
+    snapshot = history_ref.order_by_key().limit_to_last(1).get()
     # 4 hrs of history
     for key, val in snapshot.items():
-        module_logger.debug("remove child... " + val)
-        history_ref.child(key).delete()
+        if val['time'] < history_max:
+            module_logger.debug("remove child... " + val)
+            history_ref.child(key).delete()
 
 
 def internet_on():
